app/requirements.txt
streamlit==1.36.0
pandas==2.2.2
scikit-learn==1.6.1
joblib==1.5.3
psycopg2-binary==2.9.9